requirements.txt
flask
opencv-python
numpy
scipy
ultralytics
torch
netifaces
psutil
mysql-connector-python
bcrypt